Move precision_bits_from_precision to constants
