+++ olive/constants.py
@@ -95,3 +95,17 @@ class AccuracyLevel(IntEnumBase):
     fp16 = 2
     bf16 = 3
     int8 = 4
+
+
+def precision_bits_from_precision(p):
+    mapping = {
+        Precision.INT4: PrecisionBits.BITS4,
+        Precision.INT8: PrecisionBits.BITS8,
+        Precision.INT16: PrecisionBits.BITS16,
+        Precision.INT32: PrecisionBits.BITS32,
+        Precision.UINT4: PrecisionBits.BITS4,
+        Precision.UINT8: PrecisionBits.BITS8,
+        Precision.UINT16: PrecisionBits.BITS16,
+        Precision.UINT32: PrecisionBits.BITS32,
+    }
+    return mapping.get(p)
